her i et problem
