+++ ProjektMappen/MMMI/src/mmmi/UI/Main/CreateCase/CreateCaseFXMLController.java
@@ -14,6 +14,9 @@ import mmmi.UI.Main.RunFxmlLoader;
 
 import java.net.URL;
 import java.util.ResourceBundle;
+import javafx.collections.FXCollections;
+import javafx.collections.ObservableList;
+import javafx.scene.control.ScrollPane;
 
 public class CreateCaseFXMLController extends RunFxmlLoader implements Initializable {
 
@@ -28,25 +31,60 @@ public class CreateCaseFXMLController extends RunFxmlLoader implements Initializ
 
     @FXML
     private AnchorPane fxSubSubpane;
+    
+    ObservableList<String> list;
 
     /**
      * Initializes the controller class.
      */
     @Override
     public void initialize(URL url, ResourceBundle rb) {
+        
+        
+                this.list = FXCollections.observableArrayList();
+                list.add(0, "caseOpening/caseOpeningFXML.fxml");
+                list.add(1, "caseInvestigation/caseInvestigationFXML.fxml"); 
+                list.add(2, "caseOpening/caseOpeningFXML.fxml");
+                changeFxml(fxSubSubpane, list.get(0));
 
 
     }
 
     @FXML
     private void buttonHandler(ActionEvent event) {
-        if (event.getSource() == fxnextBtn) {
-            changeFxml(fxSubSubpane, "caseInvestigation/caseInvestigationFXML.fxml");
-        }
+        int count=0; 
+           if(event.getSource()==fxnextBtn){
+               if(count==0){
+                   changeFxml(fxSubSubpane, list.get(1));
+                   count = 1;
+               }
+               if(count ==1){
+                    changeFxml(fxSubSubpane, list.get(2));
+               }
+           } 
+        
+         
+  
+       
     }
 
 
 }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
