--- ProjektMappen/MMMI/src/mmmi/UI/Main/CreateCase/caseOpening/CaseOpenMainTest.java
@@ -1,40 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package mmmi.UI.Main.createCase.caseOpening;
-
-import javafx.application.Application;
-import static javafx.application.Application.launch;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.stage.Stage;
-
-/**
- * @author steff
- */
-public class CaseOpenMainTest extends Application {
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String[] args) {
-        launch(args);
-    }
-
-    @Override
-    public void start(Stage stage) throws Exception {
-        Parent root = FXMLLoader.load(getClass().getResource("caseOpeningFXML.fxml"));
-
-
-        Scene scene = new Scene(root);
-
-        stage.setScene(scene);
-        stage.setTitle("MMMI");
-        stage.show();
-    }
-
-}
-
