+++ ProjektMappen/MMMI/src/mmmi/UI/Main/CreateCase/caseOpening/caseOpeningFXML.fxml
@@ -11,7 +11,7 @@
 
 <AnchorPane fx:id="caseOpenRootpane" prefHeight="500.0" prefWidth="891.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="mmmi.UI.Main.createCase.caseOpening.CaseOpeningFXMLController">
     <children>
-        <ScrollPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="2800.0" prefWidth="891.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
+        <ScrollPane maxHeight="-Infinity" maxWidth="1.7976931348623157E308" minHeight="-Infinity" minWidth="-Infinity" prefHeight="2800.0" prefWidth="1100.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
             <content>
                 <AnchorPane fx:id="clearWhatCitizenSearchesYesCheckBox" prefHeight="2608.0" prefWidth="890.0">
                     <children>
@@ -290,4 +290,3 @@
         </ScrollPane>
     </children>
 </AnchorPane>
-
