+++ ProjektMappen/MMMI/src/mmmi/UI/Main/CreateCase/createCaseFXML.fxml
@@ -1,37 +1,27 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.SplitPane?>
-<?import javafx.scene.layout.*?>
 <?import java.net.URL?>
-<AnchorPane xmlns:fx="http://javafx.com/fxml/1" id="AnchorPane" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308"
-            prefHeight="900.0" prefWidth="1100.0" styleClass="mainFxmlClass"
-            xmlns="http://javafx.com/javafx/8.0.171" fx:controller="mmmi.UI.Main.createCase.CreateCaseFXMLController">
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.ScrollPane?>
+<?import javafx.scene.control.SplitPane?>
+<?import javafx.scene.layout.AnchorPane?>
+
+<AnchorPane id="AnchorPane" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" prefHeight="900.0" prefWidth="1100.0" styleClass="mainFxmlClass" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="mmmi.UI.Main.createCase.CreateCaseFXMLController">
     <stylesheets>
-        <URL value="@/mmmi/UI/mainStylesheet.css"/>
+        <URL value="@/mmmi/UI/mainStylesheet.css" />
     </stylesheets>
     <children>
-        <SplitPane dividerPositions="0.9173789173789174" layoutX="6.0" layoutY="6.0" orientation="VERTICAL"
-                   pickOnBounds="false" prefHeight="704.0" prefWidth="1100.0" AnchorPane.bottomAnchor="0.0"
-                   AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
+        <SplitPane dividerPositions="0.8819599109131403" layoutX="6.0" layoutY="6.0" orientation="VERTICAL" pickOnBounds="false" prefHeight="704.0" prefWidth="1100.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
             <items>
-                <AnchorPane maxWidth="1.7976931348623157E308" minHeight="0.0" minWidth="100.0" prefHeight="348.0"
-                            prefWidth="1100.0">
+            <ScrollPane maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" prefHeight="806.0" prefWidth="1098.0">
+               <content>
+                  <AnchorPane fx:id="fxSubSubpane" />
+               </content></ScrollPane>
+                <AnchorPane maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" minHeight="-Infinity" minWidth="-Infinity">
                     <children>
-                        <AnchorPane fx:id="fxSubSubpane" prefHeight="821.0" prefWidth="1098.0"
-                                    AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0"
-                                    AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0"/>
-                    </children>
-                </AnchorPane>
-                <AnchorPane maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" minHeight="-Infinity"
-                            minWidth="-Infinity">
-                    <children>
-                        <Button fx:id="fxnextBtn" layoutX="661.0" layoutY="15.0" mnemonicParsing="false"
-                                onAction="#buttonHandler" prefHeight="42.0" prefWidth="169.0" text="Næste"/>
-                        <Button fx:id="fxBackBtn" layoutX="465.0" layoutY="15.0" mnemonicParsing="false"
-                                onAction="#buttonHandler" prefHeight="42.0" prefWidth="169.0" text="Tilbage"/>
-                        <Button fx:id="fxSaveBtn" layoutX="849.0" layoutY="15.0" mnemonicParsing="false"
-                                onAction="#buttonHandler" prefHeight="42.0" prefWidth="169.0" text="Gem"/>
+                        <Button fx:id="fxnextBtn" layoutX="661.0" layoutY="15.0" mnemonicParsing="false" onAction="#buttonHandler" prefHeight="42.0" prefWidth="169.0" text="Næste" />
+                        <Button fx:id="fxBackBtn" layoutX="465.0" layoutY="15.0" mnemonicParsing="false" onAction="#buttonHandler" prefHeight="42.0" prefWidth="169.0" text="Tilbage" />
+                        <Button fx:id="fxSaveBtn" layoutX="849.0" layoutY="15.0" mnemonicParsing="false" onAction="#buttonHandler" prefHeight="42.0" prefWidth="169.0" text="Gem" />
                     </children>
                 </AnchorPane>
             </items>
